Rename and split interface.py
