+++ scenario_player/services/utils/cli/installer.py
